UCWA auth request `response_type` fixed to `token`
The `response_type` of the first OAuth request was incorrect. You can check it out yourself. I was wasting on it a lot of time...
With the old `id_token` response_type, only the `id_token` is returned. Now, with `token` resposne_type, the actual `access_token` is returned. Some of the response fields as described here also missing but I'm not sure if this is only my case.